重构 如果用户文件就让用户修改
